Copy MSDIA140.dll to target directory
Though api-ms-*.dll are copied into the build target directory (e.g.
out/Release), msdia140.dll is not. Because of this, when we set up a
new machine, we still see msdia140.dll is missing (we still have to
set PATH). Also, we see the same error to build the open source
version of goma on Win bot.

Copy msdia140 from vs_toolchains.

Change-Id: I0dee173a957f0f71a93fdefcdf1b5d87dbe1905d