if __has_include(X) returns true then X is an input
Typically __has_include is used to check if a header is available
before #include'ing it.  But in some cases __has_include is only
used as feature detection, without a subsequent #include (a real
world example is in boost's libstdcpp3.hpp).  In this case, the
goma client won't upload the file that __has_include found, and
the remote job will fail this check that succeeded locally.

What we should do add X to the list of input files if __has_include(X)
evaluates to true, but without processing X for further includes.

(And similarly for __has_include_next.)

Change-Id: Ibb77ac0691698fc965455973d7f86eb35dc646ca