add support for common multilib GCC installation directories
Clang searches for crtbegin.o files to locate candidate GCC installation
directories for header files.  The goma client quietly adds the first one
that it finds to compile jobs.  But if the selected GCC installation dir
is a multilib, like lib32 or libx32 for x86, then we also need to upload
the corresponding crtbegin.o files.  These live in "32" and "x32"
subdirectories in Ubuntu (and I believe Debian).

There are other multilib toolchains, but "lib32" is the most common one
we're likely to want to use.

To test, compare the "Candidate multilib" and "Selected multilib" lines
in clang -m32 -v between remote and local compile jobs, with the
lib32gcc-5-dev Ubuntu package installed.

Change-Id: I1d60c651f569c975891a7db92d1823ca526890db