http: Also pass the destination host in the Host field for proxied HTTP connections
This was previously being done only for HTTPS connections, and the public
commit history is not clear on why.

The previous behavior was to pass the proxy host's value in this situation,
which resulted in an HTTP connection to the proxy host that looked like
this:

    GET http://crl.pki.goog:80/GTSGIA3.crl HTTP/1.1
    Host: my-proxy.com
    ...

rather than

    GET http://crl.pki.goog:80/GTSGIA3.crl HTTP/1.1
    Host: crl.pki.goog
    ...

which, depending on the location we are connecting to, results in an error
response (crl.pki.goog returns 404 for that certificate revocation list
file, for example).

Bug: chromium:908046
Change-Id: I484f32200b15402dbccd54ad127946ab03f52fc3